refactor: adapt to typescript
