--- src/map/index.js
@@ -1,1 +1,0 @@
-export { default as tubeMap } from "./map";
