--- src/map/lines.js
@@ -1,7 +1,0 @@
-function Lines(lines) {
-    this.lines = lines;
-  }
-  
-  export default function (lines) {
-    return new Lines(lines);
-  }
